requirements.txt
PyYAML
keras
numpy
scikit-learn
tensorflow
torch
torchvision
pandas
seaborn
gensim
nltk
PyYAML
python-xmp-toolkit
num2words
ekphrasis
jupyter
jupyterlab
beautifulsoup4
html5lib
h5py
Keras
wordcloud
librosa
h5py
jupyter_contrib_nbextensions
python_speech_features
sed_eval
youtube_dl
